Don't truncate layer create parameters to 16 bits
Because they can go up to 24 bits, the C++ wrapper function didn't pass
it through correctly.

+++ src/libclient/net/message.cpp
@@ -185,7 +185,7 @@ Message makeLayerAclMessage(
 }
 
 Message makeLayerTreeCreateMessage(
-	uint8_t contextId, uint32_t id, uint16_t source, uint16_t target,
+	uint8_t contextId, uint32_t id, uint32_t source, uint32_t target,
 	uint32_t fill, uint8_t flags, const QString &name)
 {
 	QByteArray bytes = name.toUtf8();
